// react-refactor/src/types.ts
export type Player = {
  id: number;
  name: string;
  iconClass: string;
  colorClass: string;
};

export type Move = {
  player: Player;
  squareId: number;
};

export type GameStatus = {
  isComplete: boolean;
  winner: Player | undefined;
};

export type Game = {
  moves: Move[];
  status: GameStatus;
};

export type GameState = {
  currentGameMoves: Move[];
  history: {
    currentRoundGames: Game[];
    allGames: Game[];
  };
};